Revert "chore: add skeleton loading"
This reverts commit 1c85d06de20cda9ce9a905acc8cbac8c8105185a.
